Refactor: Extract HeadingsNav component and use in blog and docs
